Keep Slider along with TextInput within RangeCell
