+++ packages/components/src/mobile/bottom-sheet/range-cell.native.scss
@@ -1,10 +1,3 @@
-.sliderContainer {
-	flex: 1;
-	flex-direction: row;
-	align-content: center;
-	justify-content: space-evenly;
-}
-
 .slider {
 	flex-grow: 1;
 }
